refactor(client): move form schemas to schemas dir
