+++ apps/client/eslint.config.ts
@@ -1,6 +1,7 @@
 import config from "@clipboard-sync/eslint-config";
 import react from "eslint-plugin-react";
 import reactHooks from "eslint-plugin-react-hooks";
+// eslint-disable-next-line import/no-named-as-default
 import reactRefresh from "eslint-plugin-react-refresh";
 import globals from "globals";
