+++ apps/client/src/components/connection/form.tsx
@@ -1,4 +1,3 @@
-import { ClientNameSchema, ServerRoomIDSchema } from "@clipboard-sync/schemas";
 import {
   TextInput,
   PasswordInput,
@@ -11,17 +10,8 @@ import {
 import type { UseFormReturnType } from "@mantine/form";
 import { IconKey, IconLock, IconUser, IconX } from "@tabler/icons-react";
 import type { FC } from "react";
-import { z } from "zod";
 
-export const ConnectionFormSchema = z.object({
-  clientName: ClientNameSchema,
-  roomId: ServerRoomIDSchema,
-  secret: z.string().min(6, "Secret must be at least 6 characters"),
-  saveSecret: z.boolean(),
-  autoConnectOnStart: z.boolean(),
-});
-
-export type ConnectionFormValues = z.infer<typeof ConnectionFormSchema>;
+import type { ConnectionFormValues } from "../../schemas/connection-form";
 
 export type ConnectionFormProps = {
   form: UseFormReturnType<ConnectionFormValues>;
